Update project-responsibilities.md
An amazing VP of Engineering (DL) dropped the I for Individual from DRI and replaced it with an E for Enabler. This speaks to me enormously because too often as a DRI I felt like _I_ as the _Individual_ was responsible for all deliverables when what I am usually meant to be doing is _enabling_ deliverables to happen by my bringing the right people to the table.